refactor: extract bootstrap and rollout cli helpers
